requirements.txt
flask
pillow
numpy
matplotlib
scipy
wordcloud
